+++ deployment/layers/pooling.cpp
@@ -164,7 +164,7 @@ MaxPool2d::MaxPool2d(uint16_t input_channel_size, uint16_t input_row_size,
 
 int8_t* MaxPool2d::forward(int8_t* input, int8_t* workspace_start, uint32_t workspace_size) {
     // Getting the output start address with the input size as offset
-    int8_t* output = input == workspace_start ? workspace_start + workspace_size - ceil((this->input_channel_size * this->output_row_size * this->output_col_size) / DATA_PER_BYTE) : workspace_start;
+    int8_t* output = input == workspace_start ? workspace_start + workspace_size - (uint32_t)ceil((this->input_channel_size * this->output_row_size * this->output_col_size) / DATA_PER_BYTE) : workspace_start;
 
     int8_t temp, input_val;
 
@@ -229,7 +229,7 @@ AvgPool2d::AvgPool2d(uint16_t input_channel_size, uint16_t input_row_size,
 
 int8_t* AvgPool2d::forward(int8_t* input, int8_t* workspace_start, uint32_t workspace_size) {
     // Getting the output start address with the input size as offset
-    int8_t* output = input == workspace_start ? workspace_start + workspace_size - ceil((this->input_channel_size * this->output_row_size * this->output_col_size) / DATA_PER_BYTE) : workspace_start;
+    int8_t* output = input == workspace_start ? workspace_start + workspace_size - (uint32_t)ceil((this->input_channel_size * this->output_row_size * this->output_col_size) / DATA_PER_BYTE) : workspace_start;
 
     int16_t total;
